Added type definition file
