+++ src/types.ts
@@ -1,0 +1,6 @@
+export interface TokenDetails {
+  address: string;
+  decimals: number;
+  symbol: string;
+  name: string;
+}
